requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
transformers
requests
python-dotenv